Use a Conditional object pattern for platforms
We have methods that execute a PlatformAction given some condition, but
they're a grab bag: onRpm, onWindows, etc. By refactoring to use a
Conditional pattern, we can enforce that our tests are explicitly
accounting for every option for a particular condition. This change also
groups the code a little more logically, so it's easier to see what the
potential platforms, service managers, and package managers are.