Allow test writers to prefer deprecated paths
There is a strange edge case in our tests where a rest API might be
defined with a mix of deprecated and nondeprecated paths that share the
same parameters. In that case, warnings would return inconsistently.
This could be mitigated with the "allowed_warnings" list for a "do"
section in the rest test definition, but that does not allow a developer
to explicitly test the deprecated version of the path and verify that it
returns the correct warning, nor verify that the non-deprecated version
never returns the warning.

This commit aims to fix that by introducing a
"prefer_non_deprecated_api_paths" flag for the "do" section that
defaults to "true" and will make sure that paths are chosen with
consistent deprecation status.